+++ code/win32_file_io.h
@@ -31,7 +31,7 @@
 //
 // File reading using the win32 API. Reads the entire file into memory that Data is pointing to.
 // Uses malloc, you will need to free Data.
-b32 win32_ReadFile(char const *PathAndName, u8 **Data, size_t *DataSize);
+b32 win32_ReadFile(char const *PathAndName, u8 **Data, u32 *DataSize);
